Finish part 4 (#1)
* Finish 4.20

* Finish 4.21

* Finish 4.11

* Fix failed blog making tests

Also create login api tests

* Partial test bugfixes

* Fixing deleting blogs tests

* Finish 4.23

+++ part4/bloglist/controllers/blogsRouter.js
@@ -1,45 +1,62 @@
 import { Router } from 'express';
 // eslint-disable-next-line import/extensions
 import blogRepository from '../repositories/BlogRepositorySingleton.js';
-import authService from '../services/AuthService.js';
-import userRepository from '../repositories/UserRepositorySingleton.js';
 
 const blogsRouter = new Router();
 
-blogsRouter.get('/', async (request, response) => {
-    response.json(await blogRepository.getAllBlogs());
+blogsRouter.get('/:id', async (req, res) => {
+    const blog = await blogRepository.getBlogById(req.params.id.toString());
+    if (!blog) {
+        res.status(404)
+            .send({ error: 'blog not found' });
+    }
+
+    res.status(200)
+        .send(blog);
 });
 
 blogsRouter.post('/', async (request, response) => {
-    const requestToken = authService.getToken(request.get('authorization'));
-    const decodedToken = authService.decodeToken(requestToken);
+    const { user } = request;
 
-    if (!decodedToken.id) {
-        response.status(401)
+    if (!user) {
+        return response.status(401)
             .json({ error: 'token invalid' });
     }
 
-    const user = await userRepository.getUserById(decodedToken.id);
     const newBlog = {
         ...request.body,
-        // eslint-disable-next-line no-underscore-dangle
-        user: user._id,
+        user,
     };
 
     const result = await blogRepository.addNewBlog(newBlog);
-    response.status(201)
+    return response.status(201)
         .json(result);
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-    const id = request.params.id.toString();
-    const result = await blogRepository.deleteBlogById(id);
+    const { user } = request;
 
-    if (!result) {
-        throw new Error('blog not found');
+    if (!user) {
+        return response.status(401)
+            .json({ error: 'token invalid' });
     }
 
-    response.status(204)
+    const id = request.params.id.toString();
+    const toDeleteBlog = await blogRepository.getBlogById(id);
+
+    if (!toDeleteBlog) {
+        return response.status(404)
+            .send({ error: 'blog not found' });
+    }
+
+    if (toDeleteBlog.user.toString() !== user) {
+        return response.status(401)
+            .send({ error: 'wrong user' });
+    }
+
+    await blogRepository.deleteBlogById(id);
+
+    return response.status(204)
         .end();
 });
 
@@ -50,7 +67,8 @@ blogsRouter.put('/:id', async (request, response) => {
     const result = await blogRepository.updateBlogById(id, blog);
 
     if (!result) {
-        throw new Error('blog not found');
+        return response.status(404)
+            .send({ error: 'blog not found' });
     }
 
     response.status(204)
